introduce decimal sells into our pos

+++ src/validators/admin/inventory-validator.ts
@@ -3,7 +3,7 @@ import { z } from "zod";
 export const addStockSchema = z.object({
   shopId: z.string().min(1),
   productId: z.string().min(1),
-  quantity: z.coerce.number().int().positive(),
+  quantity: z.coerce.number().positive(),
   costPrice: z.coerce.number().nonnegative(),
   sellingPrice: z.coerce.number().positive(),
   reorderLevel: z.coerce.number().int().nonnegative(),
@@ -15,7 +15,7 @@ export const addStockSchema = z.object({
 
 export const adjustStockSchema = z.object({
   inventoryId: z.string().min(1),
-  quantity: z.coerce.number().int(),
+  quantity: z.coerce.number(),
   reason: z.string().trim().min(3).max(500),
 });
